backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication
django-allauth==0.63.3
dj-rest-auth==6.0.0
PyJWT>=2.8.0
cryptography>=41.0.0

# Environment
python-dotenv==1.0.1

# HTTP Client for Asaas
httpx==0.27.0
requests>=2.31.0

# Database
psycopg2-binary==2.9.9

# Utilities
python-decouple==3.8
Pillow>=11.0.0
